refactor bulk payments with submission controller and bulk payment controller
