+++ app/controllers/events_controller.rb
@@ -2,7 +2,7 @@ class EventsController < ApplicationController
   include AhoyTracking, TagAssignable
   skip_before_action :authenticate_user!, only: [ :index, :show, :staff ]
   skip_before_action :verify_authenticity_token, only: [ :preview ]
-  before_action :set_event, only: %i[ show edit update destroy preview dashboard sample_ticket background registrants onboarding staff edit_staff update_staff recipients bulk_payments preview_reminder confirm_reminder send_reminder copy_registration_form allocate_bulk_payment create_bulk_payment ]
+  before_action :set_event, only: %i[ show edit update destroy preview dashboard sample_ticket background registrants onboarding staff edit_staff update_staff recipients preview_reminder confirm_reminder send_reminder copy_registration_form ]
 
   def index
     authorize!
@@ -213,99 +213,6 @@ class EventsController < ApplicationController
     @dashboard = EventDashboard.new(@event)
   end
 
-  def bulk_payments
-    authorize! @event
-
-    @event = @event.decorate
-    # Shared across every card so attendee matching and allocated totals don't
-    # re-query per registration per card.
-    @event_registrations = @event.event_registrations.active.includes(:registrant)
-    @submissions = @event.form_submissions
-                         .where(role: "bulk_payment")
-                         .includes(:person, form_answers: :form_field, payment: :allocations)
-                         .order(created_at: :desc)
-    @allocated_by_registration = allocated_cents_by_registration(@event_registrations)
-  end
-
-  def allocate_bulk_payment
-    authorize! @event
-    @event = @event.decorate
-    payment = Payment.find(params[:payment_id])
-    event_registration = EventRegistration.find_by(id: params[:event_registration_id])
-    unless event_registration
-      flash.now[:alert] = "Please select a registrant"
-      assign_allocation_card_data(payment)
-      respond_to do |format|
-        format.turbo_stream
-        format.html { redirect_to bulk_payments_event_path(@event), alert: "Please select a registrant" }
-      end
-      return
-    end
-    amount_cents = (params[:amount_dollars].to_d * 100).to_i
-
-    if amount_cents <= 0
-      flash.now[:alert] = "Amount must be greater than $0.00"
-    elsif amount_cents > (payment.amount_cents_remaining || 0)
-      flash.now[:alert] = "Amount exceeds remaining balance"
-    else
-      allocation = Allocation.new(source: payment, allocatable: event_registration, amount: amount_cents)
-      if allocation.save
-        flash.now[:notice] = "Allocation successful"
-      else
-        flash.now[:alert] = allocation.errors.full_messages.to_sentence
-      end
-    end
-
-    assign_allocation_card_data(payment)
-
-    respond_to do |format|
-      format.turbo_stream
-      format.html { redirect_to bulk_payments_event_path(@event), notice: flash.now[:alert] || "Allocation successful" }
-    end
-  end
-
-  def create_bulk_payment
-    authorize! @event
-    @event = @event.decorate
-    @event_registrations = @event.event_registrations.active.includes(:registrant)
-    @allocated_by_registration = allocated_cents_by_registration(@event_registrations)
-
-    submission = @event.form_submissions.find(params[:submission_id])
-    payment_type = params[:payment_type]
-
-    unless %w[CashPayment CheckPayment].include?(payment_type)
-      flash.now[:alert] = "Invalid payment type"
-      respond_to do |format|
-        format.turbo_stream
-        format.html { redirect_to bulk_payments_event_path(@event), alert: "Invalid payment type" }
-      end
-      return
-    end
-
-    payment = submission.build_payment(
-      amount_cents: (params[:amount_dollars].to_d * 100).to_i,
-      currency: params[:currency].presence || "usd",
-      type: payment_type,
-      check_number: params[:check_number].presence,
-      memo: params[:memo].presence
-    )
-    payment.payer_sgid = params[:payer_sgid]
-    payment.additional_designation_sgid = params[:additional_designation_sgid]
-
-    if payment.save
-      @payment = payment
-      @submission = submission.decorate
-      flash.now[:notice] = "Payment recorded"
-    else
-      flash.now[:alert] = payment.errors.full_messages.to_sentence
-    end
-
-    respond_to do |format|
-      format.turbo_stream
-      format.html { redirect_to bulk_payments_event_path(@event), notice: flash.now[:alert] || "Payment recorded" }
-    end
-  end
-
   def preview_reminder
     authorize! @event
     @event = @event.decorate
@@ -514,25 +421,6 @@ class EventsController < ApplicationController
       .where(id: allowed_ids)
       .includes(registrant: [ :user, :contact_methods ])
       .select { |r| r.registrant.preferred_email.present? }
-  end
-
-  # Reloads the payment and the data its bulk payment card needs, so the
-  # allocate turbo stream can re-render the whole card with fresh due/allocated
-  # totals and re-evaluate whether each registration is now paid in full.
-  def assign_allocation_card_data(payment)
-    @payment = payment.reload
-    @submission = @payment.form_submission
-    @event_registrations = @event.event_registrations.active.includes(:registrant)
-    @allocated_by_registration = allocated_cents_by_registration(@event_registrations)
-  end
-
-  # Allocated cents per registration id, fetched in one grouped query so the
-  # bulk payment cards read totals from a hash instead of querying per row.
-  def allocated_cents_by_registration(registrations)
-    Allocation
-      .where(allocatable_type: "EventRegistration", allocatable_id: registrations.ids)
-      .group(:allocatable_id)
-      .sum(:amount)
   end
 
   # Maps registrant person_id => the organization name they typed on the
