+++ app/views/events/bulk_payment_submissions/show.html.erb
