+++ app/views/events/bulk_payment_submissions/ticket.html.erb
