+++ app/views/events/bulk_payments/allocate.turbo_stream.erb
