+++ app/views/events/bulk_payments/create.turbo_stream.erb
@@ -2,7 +2,7 @@
 
 <% unless flash[:alert] %>
   <%= turbo_stream.replace "payment-card-#{@submission.id}" do %>
-    <%= render "bulk_payment_card", submission: @submission,
+    <%= render "events/bulk_payment_card", submission: @submission,
           event_registrations: @event_registrations,
           allocated_by_registration: @allocated_by_registration,
           expanded: true %>
