+++ app/views/events/bulk_payments/unlink.turbo_stream.erb
@@ -1,0 +1,10 @@
+<%= turbo_stream.replace "flash_now", partial: "shared/flash_messages" %>
+
+<% if @submission %>
+  <%= turbo_stream.replace "payment-card-#{@submission.id}" do %>
+    <%= render "events/bulk_payment_card", submission: @submission,
+          event_registrations: @event_registrations,
+          allocated_by_registration: @allocated_by_registration,
+          expanded: true %>
+  <% end %>
+<% end %>
