+++ db/migrate/20260727000000_add_metadata_to_form_submissions.rb
@@ -1,0 +1,5 @@
+class AddMetadataToFormSubmissions < ActiveRecord::Migration[8.1]
+  def change
+    add_column :form_submissions, :metadata, :json
+  end
+end
