+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_07_15_053137) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_27_000000) do
   create_table "action_text_mentions", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.bigint "action_text_rich_text_id", null: false
     t.datetime "created_at", null: false
@@ -654,6 +654,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_07_15_053137) do
     t.datetime "created_at", null: false
     t.bigint "event_id"
     t.integer "form_id", null: false
+    t.json "metadata"
     t.bigint "person_id", null: false
     t.string "role"
     t.string "slug"
