+++ spec/requests/events_spec.rb
@@ -2279,235 +2279,6 @@ RSpec.describe "Events", type: :request do
     end
   end
 
-  describe "GET /events/:id/bulk_payments" do
-    let(:admin) { create(:user, :admin) }
-    let(:event) { create(:event, cost_cents: 2500) }
-    let(:bulk_form) { create(:form) }
-    let!(:event_form) { create(:event_form, event: event, form: bulk_form, role: "bulk_payment") }
-    let(:payer) { create(:person) }
-    let!(:submission) { create(:form_submission, person: payer, form: bulk_form, event: event, role: "bulk_payment") }
-    let!(:attendees_field) do
-      create(:form_field, form: bulk_form, field_identifier: "number_of_attendees", name: "Attendees")
-    end
-
-    before { sign_in admin }
-
-    it "shows the submitted amount even when no payment has landed" do
-      submission.form_answers.create!(form_field: attendees_field, submitted_answer: "3")
-
-      get bulk_payments_event_path(event)
-
-      expect(response).to have_http_status(:ok)
-      expect(response.body).to include("$75")
-    end
-
-    it "shows the recorded payment amount when a payment exists" do
-      submission.form_answers.create!(form_field: attendees_field, submitted_answer: "3")
-      create(:payment, person: payer, form_submission: submission,
-             amount_cents: 5000, amount_cents_remaining: 5000)
-
-      get bulk_payments_event_path(event)
-
-      expect(response.body).to include("$50")
-    end
-
-    it "renders the targeted submission's row expanded when given an expand param" do
-      get bulk_payments_event_path(event, expand: submission.id)
-
-      expect(response.body).to include("id=\"payment-card-#{submission.id}\"")
-      # Expanded server-side: the toggle button gets data-dropdown-target="expand"
-      # so the dropdown controller clicks it on connect to open the card.
-      expect(response.body).to include("data-dropdown-target=\"expand\"")
-    end
-
-    it "renders rows collapsed without an expand param" do
-      get bulk_payments_event_path(event)
-
-      # Collapsed: the details panel keeps the `hidden` class and this card's chevron is not rotated.
-      expect(response.body).to match(/id="payment-details-#{submission.id}"\s+class="hidden/)
-      expect(response.body).not_to match(/id="payment-arrow-#{submission.id}"[^>]*rotate-180/)
-    end
-
-    it "renders a Profile column with a circle-only profile button for matched attendees" do
-      attendee = create(:person, first_name: "Match", last_name: "Attendee", email: "match.attendee@example.com")
-      create(:event_registration, event: event, registrant: attendee, status: "registered")
-      create(:form_field, form: bulk_form, field_identifier: "bulk_payment_attendees", name: "Attendees list")
-      submission.form_answers.create!(
-        form_field: bulk_form.form_fields.find_by(field_identifier: "bulk_payment_attendees"),
-        submitted_answer: [ { first_name: "Match", last_name: "Attendee", email: "match.attendee@example.com" } ].to_json
-      )
-
-      get bulk_payments_event_path(event)
-
-      expect(response).to have_http_status(:ok)
-      expect(response.body).to include(">Profile<")
-      # Circle-only profile button: a boxed (sky) link to the person holding just
-      # the compact h-5 avatar circle.
-      expect(response.body).to include(person_path(attendee))
-      expect(response.body).to include("bg-sky-100")
-      expect(response.body).to include("h-5 w-5")
-    end
-
-    it "shows a grey \"Paid\" instead of an orange balance when the registration is fully covered" do
-      attendee = create(:person, first_name: "Paid", last_name: "Infull", email: "paid.infull@example.com")
-      registration = create(:event_registration, event: event, registrant: attendee, status: "registered")
-      create(:form_field, form: bulk_form, field_identifier: "bulk_payment_attendees", name: "Attendees list")
-      submission.form_answers.create!(
-        form_field: bulk_form.form_fields.find_by(field_identifier: "bulk_payment_attendees"),
-        submitted_answer: [ { first_name: "Paid", last_name: "Infull", email: "paid.infull@example.com" } ].to_json
-      )
-      # Fully cover the $25 registration fee so nothing is owed.
-      create(:allocation, source: create(:payment, amount_cents: 2500, amount_cents_remaining: 2500),
-             allocatable: registration, amount: 2500)
-
-      get bulk_payments_event_path(event)
-
-      expect(response).to have_http_status(:ok)
-      expect(response.body).to include("text-gray-500 whitespace-nowrap\">Paid<")
-      expect(response.body).not_to include("$0.00")
-    end
-
-    it "lists only registrants who are not yet fully paid in the new-allocation dropdown" do
-      # The dropdown only renders when the payment still has an unallocated balance.
-      create(:payment, person: payer, form_submission: submission,
-             amount_cents: 5000, amount_cents_remaining: 5000)
-      unpaid = create(:person, first_name: "Owes", last_name: "Money", email: "owes.money@example.com")
-      create(:event_registration, event: event, registrant: unpaid, status: "registered")
-      paid = create(:person, first_name: "All", last_name: "Square", email: "all.square@example.com")
-      paid_registration = create(:event_registration, event: event, registrant: paid, status: "registered")
-      # Fully cover the $25 registration fee so this registrant drops off the list.
-      create(:allocation, source: create(:payment, amount_cents: 2500, amount_cents_remaining: 2500),
-             allocatable: paid_registration, amount: 2500)
-
-      get bulk_payments_event_path(event)
-
-      expect(response).to have_http_status(:ok)
-      expect(response.body).to include("Owes Money — owes.money@example.com")
-      expect(response.body).not_to include("All Square — all.square@example.com")
-    end
-  end
-
-  describe "POST /events/:id/allocate_bulk_payment" do
-    let(:admin) { create(:user, :admin) }
-    let(:event) { create(:event) }
-    let(:bulk_form) { create(:form) }
-    let!(:event_form) { create(:event_form, event: event, form: bulk_form, role: "bulk_payment") }
-    let(:payer) { create(:person) }
-    let!(:submission) { create(:form_submission, person: payer, form: bulk_form, event: event, role: "bulk_payment") }
-    let!(:payment) { create(:payment, person: payer, form_submission: submission,
-                            amount_cents: 1000, amount_cents_remaining: 1000) }
-    let(:registrant) { create(:person) }
-    let!(:event_registration) { create(:event_registration, event: event, registrant: registrant) }
-
-    let(:valid_params) do
-      { payment_id: payment.id, event_registration_id: event_registration.id, amount_dollars: "5.00" }
-    end
-
-    before { sign_in admin }
-
-    it "rejects unauthenticated request" do
-      sign_out admin
-      post allocate_bulk_payment_event_path(event), params: valid_params
-      expect(response).to redirect_to(new_user_session_path)
-    end
-
-    it "rejects non-admin request" do
-      sign_out admin
-      sign_in create(:user)
-      post allocate_bulk_payment_event_path(event), params: valid_params
-      expect(response).to redirect_to(root_path)
-    end
-
-    it "creates an allocation and returns turbo_stream" do
-      expect {
-        post allocate_bulk_payment_event_path(event), params: valid_params, as: :turbo_stream
-      }.to change(Allocation, :count).by(1)
-
-      expect(payment.reload.amount_cents_remaining).to eq(500)
-      expect(response.media_type).to eq(Mime[:turbo_stream])
-      expect(response.body).to include("Allocation successful")
-    end
-
-    context "when the allocation pays a matched registration in full" do
-      let(:event) { create(:event, cost_cents: 500) }
-      let(:registrant) { create(:person, email: "match@example.com") }
-      let!(:attendees_field) do
-        create(:form_field, form: bulk_form, field_identifier: "bulk_payment_attendees", name: "Attendees")
-      end
-
-      before do
-        submission.form_answers.create!(
-          form_field: attendees_field,
-          submitted_answer: [ { first_name: registrant.first_name, last_name: registrant.last_name, email: "match@example.com" } ].to_json
-        )
-      end
-
-      it "re-renders the card with refreshed totals and hides the inline allocate box for that registration" do
-        post allocate_bulk_payment_event_path(event),
-             params: { payment_id: payment.id, event_registration_id: event_registration.id, amount_dollars: "5.00" },
-             as: :turbo_stream
-
-        # The whole card is re-rendered (kept expanded) with the new totals: the
-        # registration is now fully paid, so its due button reads "Paid".
-        expect(response.body).to include("payment-card-#{submission.id}")
-        expect(response.body).to include("rotate-180")
-        expect(response.body).to include(">Paid</span>")
-        # Payment still has $5 left, so only the bottom "New allocation" form
-        # remains — the inline per-row box for the now-paid registration is gone.
-        expect(response.body.scan(">Allocate</button>").size).to eq(1)
-      end
-    end
-
-    it "shows alert when event_registration_id is blank" do
-      params = valid_params.merge(event_registration_id: "")
-      expect {
-        post allocate_bulk_payment_event_path(event), params: params, as: :turbo_stream
-      }.not_to change(Allocation, :count)
-
-      expect(response.body).to include("Please select a registrant")
-    end
-
-    it "shows alert when event_registration_id is invalid" do
-      params = valid_params.merge(event_registration_id: 999999)
-      expect {
-        post allocate_bulk_payment_event_path(event), params: params, as: :turbo_stream
-      }.not_to change(Allocation, :count)
-
-      expect(response.body).to include("Please select a registrant")
-    end
-
-    it "shows alert when amount is zero" do
-      params = valid_params.merge(amount_dollars: "0.00")
-      expect {
-        post allocate_bulk_payment_event_path(event), params: params, as: :turbo_stream
-      }.not_to change(Allocation, :count)
-
-      expect(response.body).to include("Amount must be greater than $0.00")
-    end
-
-    it "shows alert when amount exceeds remaining balance" do
-      params = valid_params.merge(amount_dollars: "20.00")
-      expect {
-        post allocate_bulk_payment_event_path(event), params: params, as: :turbo_stream
-      }.not_to change(Allocation, :count)
-
-      expect(response.body).to include("Amount exceeds remaining balance")
-    end
-
-    it "shows alert when event registration is already fully paid" do
-      large_payment = create(:payment, person: payer, form_submission: submission,
-                             amount_cents: 2000, amount_cents_remaining: 2000)
-      create(:allocation, source: large_payment, allocatable: event_registration, amount: 1099)
-
-      params = { payment_id: large_payment.id, event_registration_id: event_registration.id, amount_dollars: "5.00" }
-      expect {
-        post allocate_bulk_payment_event_path(event), params: params, as: :turbo_stream
-      }.not_to change(Allocation, :count)
-
-      expect(response.body).to include("already fully paid")
-    end
-  end
-
   describe "POST /send_reminder" do
     let!(:registration_one) { create(:event_registration, event: event) }
     let!(:registration_two) { create(:event_registration, event: event) }
